Introduce IAvailableDataRepositry and refactor charge link receipt to use it
